+++ src/handler.ts
@@ -1,6 +1,6 @@
 import { APIGatewayEvent, APIGatewayProxyHandler, Context } from 'aws-lambda';
 import 'source-map-support/register';
-import { createItemInDB, getItemFromDB, getAllItemsFromDB, deleteItemFromDB } from "./dynamodb-actions";
+import { createItemInDB, getItemFromDB, getAllItemsFromDB, deleteItemFromDB, updateItemInDB } from "./dynamodb-actions";
 
 /** Save an item in the to-do list */
 export const createToDoItem: APIGatewayProxyHandler = async (
@@ -8,13 +8,33 @@ export const createToDoItem: APIGatewayProxyHandler = async (
   context: Context) => {
 
   console.log('event.body >>', event.body);
-  const incoming: { item: string; complete: boolean } = JSON.parse(event.body);
-  const { item, complete } = incoming;
+  const incoming: { todoItem: string; complete: boolean } = JSON.parse(event.body);
+  const { todoItem, complete } = incoming;
 
   try {
-    await createItemInDB(item, complete);
+    await createItemInDB(todoItem, complete);
 
     return buildResponse({ created: incoming }, 201);
+  } catch (err) {
+    console.log("Error: ", err);
+    return buildResponse(err, 400);
+  }
+};
+
+export const updateToDoItem: APIGatewayProxyHandler = async (
+  event: APIGatewayEvent,
+  context: Context) => {
+
+  const id: string = event.pathParameters.id;  
+  console.log('updated id >>', id);
+  console.log('updated event.body >>', event.body);
+  const incoming: { todoItem: string; complete: boolean } = JSON.parse(event.body);
+  const { todoItem, complete } = incoming;
+
+  try {
+    const result = await updateItemInDB(id, todoItem, complete);
+    console.log('updated result >>', result);
+    return buildResponse({ updated: incoming }, 201);
   } catch (err) {
     console.log("Error: ", err);
     return buildResponse(err, 400);
@@ -27,9 +47,9 @@ export const getToDoItem: APIGatewayProxyHandler = async (
     const id: string = event.pathParameters.id;
 
     try {
-      const toDoItem = await getItemFromDB(id);
+      const response = await getItemFromDB(id);
   
-      return buildResponse(toDoItem, 200);
+      return buildResponse(response, 200);
     } catch (err) {
       console.log("Error: ", err);
       return buildResponse(err, 404);
@@ -41,9 +61,9 @@ export const getAllToDoItems: APIGatewayProxyHandler = async (
   context: Context) => {
 
     try {
-      const toDoItems = await getAllItemsFromDB();
+      const response = await getAllItemsFromDB();
   
-      return buildResponse(toDoItems, 200);
+      return buildResponse(response, 200);
     } catch (err) {
       console.log("Error: ", err);
       return buildResponse(err, 404);